feat(EncryptionContactChip): Password and chips are correctly set when opening a draft created from the web
