+++ app/src/main/res/layout/chip_encryption_contact.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+  ~ Infomaniak Mail - Android
+  ~ Copyright (C) 2025 Infomaniak Network SA
+  ~
+  ~ This program is free software: you can redistribute it and/or modify
+  ~ it under the terms of the GNU General Public License as published by
+  ~ the Free Software Foundation, either version 3 of the License, or
+  ~ (at your option) any later version.
+  ~
+  ~ This program is distributed in the hope that it will be useful,
+  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of
+  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  ~ GNU General Public License for more details.
+  ~
+  ~ You should have received a copy of the GNU General Public License
+  ~ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+  -->
+
+<com.infomaniak.mail.ui.newMessage.BackspaceAwareChip xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    style="@style/EncryptionEmailAddressChip"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_marginHorizontal="2dp"
+    android:ellipsize="end"
+    android:focusableInTouchMode="true"
+    app:chipIconSize="@dimen/mediumIconSize"
+    app:chipIconTint="@color/encryptionIconColor"
+    app:chipMinTouchTargetSize="44dp"
+    app:iconStartPadding="@dimen/marginStandardVerySmall"
+    tools:chipIcon="@drawable/ic_lock_filled"
+    tools:text="@tools:sample/full_names" />
